Move server-side auth user logic from _app.js to a HOC to support static HTML rendering
