+++ examples/with-firebase-authentication-serverless/pages/example.js
@@ -4,7 +4,10 @@ import Link from 'next/link'
 const Example = props => {
   return (
     <div>
-      <p>This page does not fetch any data or include the authed user info.</p>
+      <p>
+        This page is static because it does not fetch any data or include the
+        authed user info.
+      </p>
       <Link href={'/'}>
         <a>Home</a>
       </Link>
